施工能耗两个新页面

+++ Gksyb.Server/Services/Services/Dto/GetConstructionEchartDataResponse.cs
@@ -8,7 +8,7 @@ namespace Gksyb.Server.Services.Services.Dto
 {
     public class GetConstructionEchartDataResponse
     {
-        public List<int> FreshWaterCostList { get; set; }
+        public List<decimal> FreshWaterCostList { get; set; }
         public List<decimal> DieselOilCostList { get; set; }
         public List<decimal> LubeCostList { get; set; }
     }
